Moving all BMS functions to a seperate driver
