+++ BMS Workspace-git/bms-1227-fw/include/phantom_can.h
@@ -11,7 +11,7 @@
 #include "stdint.h"
 
 void CANSend(uint8_t can_msg[]);
-void getVCUMSG(void);
+char getVCUMSG(void);
 
 
 #endif /* INCLUDE_PHANTOM_CAN_H_ */
